Addition of modules file
